Make token class header-only
