+++ include/booleval/token/token.h
@@ -47,16 +47,31 @@ namespace token {
  */
 class token {
 public:
-    token() noexcept;
+    token() noexcept
+        : type_(token_type::unknown)
+    {}
+
     token(token&& rhs) = default;
     token(token const& rhs) = default;
-    token(token_type const type) noexcept;
-    token(std::string_view const value) noexcept;
-    token(token_type const type, std::string_view const value) noexcept;
+
+    token(token_type const type) noexcept
+        : type_(type), value_(map_to_token_value(type))
+    {}
+
+    token(std::string_view const value) noexcept
+        : type_(map_to_token_type(value)), value_(value)
+    {}
+
+    token(token_type const type, std::string_view const value) noexcept
+        : type_(type), value_(value)
+    {}
 
     token& operator=(token&& rhs) = default;
     token& operator=(token const& rhs) = default;
-    bool operator==(token const& rhs) const noexcept;
+
+    [[nodiscard]] bool operator==(token const& rhs) const noexcept {
+        return type_  == rhs.type_ && value_ == rhs.value_;
+    }
 
     ~token() = default;
 
@@ -65,28 +80,40 @@ public:
      *
      * @param type Token type
      */
-    void type(token_type const type) noexcept;
+    void type(token_type const type) noexcept {
+        if (type != type_) {
+            type_ = type;
+            value_ = map_to_token_value(type);
+        }
+    }
 
     /**
      * Gets the token type.
      *
      * @return Token type
      */
-    token_type type() const noexcept;
+    token_type type() const noexcept {
+        return type_;
+    }
 
     /**
      * Sets the token value.
      *
      * @param value Token value
      */
-    void value(std::string_view const value) noexcept;
+    void value(std::string_view const value) noexcept {
+        type_ = map_to_token_type(value);
+        value_ = value;
+    }
 
     /**
      * Gets the token value.
      *
      * @return Token value
      */
-    [[nodiscard]] std::string_view value() const noexcept;
+    [[nodiscard]] std::string_view value() const noexcept {
+        return value_;
+    }
 
     /**
      * Gets the token value as an arithmetic value.
@@ -96,28 +123,36 @@ public:
      */
     template <typename T,
               typename = std::enable_if_t<std::is_arithmetic_v<T>>>
-    [[nodiscard]] std::optional<T> value() const noexcept;
+    [[nodiscard]] std::optional<T> value() const noexcept {
+        return utils::from_chars<T>(value_);
+    }
 
     /**
      * Checks whether the token is of the specified type.
      *
      * @return True if the token is of the specified type, false otherwise
      */
-    [[nodiscard]] bool is(token_type const type) const noexcept;
+    [[nodiscard]] bool is(token_type const type) const noexcept {
+        return type_ == type;
+    }
 
     /**
      * Checks whether the token is not of the specified type.
      *
      * @return True if the token is not of the specified type, false otherwise
      */
-    [[nodiscard]] bool is_not(token_type const type) const noexcept;
+    [[nodiscard]] bool is_not(token_type const type) const noexcept {
+        return type_ != type;
+    }
 
     /**
      * Checks whether the token is of the first or second specified type.
      *
      * @return True if the token is of the first or second specified type, false otherwise
      */
-    [[nodiscard]] bool is_one_of(token_type const first, token_type const second) const noexcept;
+    [[nodiscard]] bool is_one_of(token_type const first, token_type const second) const noexcept {
+        return is(first) || is(second);
+    }
 
     /**
      * Checks whether the token is one of the multiple specified types.
@@ -125,22 +160,16 @@ public:
      * @return True if the token is one of the multiple specified types, false otherwise
      */
     template <typename... TokenType>
-    [[nodiscard]] bool is_one_of(token_type const first, token_type const second, TokenType const ... nth) const noexcept;
+    [[nodiscard]] bool is_one_of(token_type const first,
+                                 token_type const second,
+                                 TokenType const ... nth) const noexcept {
+        return is(first) || is_one_of(second, nth...);
+    }
 
 private:
     token_type type_;
     std::string_view value_;
 };
-
-template <typename T, typename>
-std::optional<T> token::value() const noexcept {
-    return utils::from_chars<T>(value_);
-}
-
-template <typename... TokenType>
-bool token::is_one_of(token_type const first, token_type const second, TokenType const ... nth) const noexcept {
-    return is(first) || is_one_of(second, nth...);
-}
 
 } // token
